refactor Command and actionJoin
