+++ src/common/magicNumber.h
@@ -6,6 +6,7 @@
 #define SUPPORTED_CHANNEL_MODES "kost"
 #define CHANNEL_PREFIXES "#&"
 #define CHANNEL_USER_LIMIT_DISABLED -1
+#define MAX_JOIN_CHANNELS 20
 #define SUCCESS 0
 #define FAILURE 1
 #define EXIT_SUCCESS 0
